Catalog rule implementation complete, new product price mechanish remaining
